pusher_wishlist

+++ app/Http/Controllers/client/WishListController.php
@@ -8,6 +8,7 @@ use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Log;
 
 class WishListController extends Controller
 {
@@ -18,39 +19,55 @@ class WishListController extends Controller
 
         return view('client.wish-list', compact('wishlists'));
     }
-
     public function addToWishlist(Request $request)
     {
         $userId = Auth::id();
         $existingWishlistItem = Wishlist::where('user_id', $userId)
             ->where('product_id', $request->product_id)
             ->first();
-
+    
         if ($existingWishlistItem) {
+           // Log::info("Sản phẩm đã có trong wishlist: " . $request->product_id);
             return redirect()->back()->with('message', 'Sản phẩm đã có trong wishlist.');
         }
-
+    
         $wishlist = new Wishlist();
         $wishlist->user_id = $userId;
         $wishlist->product_id = $request->product_id;
         $wishlist->save();
-
-        event(new ProductAddedToWishlist($userId, $request->product_id));
-
-        return redirect()->route('list-wish')->with('message', 'Sản phẩm được thêm vào wishlist thành công.');
+    
+        $product = Product::where('product_id', $request->product_id)->firstOrFail();
+        $productName = $product->product_name;
+    
+        $user = Auth::user(); 
+        $username = $user->username;
+    
+        event(new ProductAddedToWishlist($userId, $username, $productName, $request->product_id, 'add'));
+    
+        return redirect()->back()->with('message', 'Sản phẩm được thêm vào wishlist thành công.');
     }
-
-
+    
     public function deleteWishlist($id)
     {
         $userId = Auth::id();
-
+    
         $wishlistItem = Wishlist::where('user_id', $userId)
             ->where('product_id', $id)
             ->first();
-
-        $wishlistItem->delete();
-
-        return redirect()->back()->with('message', 'Sản phẩm đã bị xóa khỏi wishlist.');
+    
+        if ($wishlistItem) {
+            $productName = $wishlistItem->product->product_name;
+            $wishlistItem->delete();
+    
+            $user = Auth::user();
+            $username = $user->username;
+    
+            event(new ProductAddedToWishlist($userId, $username, $productName, $id, 'remove'));
+            
+            return redirect()->back()->with('message', 'Sản phẩm đã bị xóa khỏi wishlist.');
+        }
+    
+        return redirect()->back()->with('message', 'Sản phẩm không tồn tại trong wishlist.');
     }
+    
 }
